requirements.txt
flask>=3.0
requests>=2.31
supabase>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
